+++ bin/publish-release.sh
@@ -34,7 +34,11 @@ fi
 # return to branch
 git checkout --quiet "$current_branch"
 
-if [[ "$success" = "1" ]]
+# restore generated dist files
+git checkout --quiet "v$version" -- dist
+git reset --quiet -- dist
+
+if [[ "$success" == "1" ]]
 then
 	echo "Success."
 else