+++ src/list/ListView.js
@@ -76,21 +76,36 @@ var ListViewGrid = Grid.extend({
 	// slices by day
 	spanToSegs: function(span) {
 		var view = this.view;
-		var dayStart = view.start.clone();
-		var dayEnd;
+		var dayStart = view.start.clone().time(0); // timed, so segs get times!
+		var dayIndex = 0;
 		var seg;
 		var segs = [];
 
 		while (dayStart < view.end) {
-			dayEnd = dayStart.clone().add(1, 'day');
+
 			seg = intersectRanges(span, {
 				start: dayStart,
-				end: dayEnd
+				end: dayStart.clone().add(1, 'day')
 			});
+
 			if (seg) {
+				seg.dayIndex = dayIndex;
 				segs.push(seg);
 			}
-			dayStart = dayEnd;
+
+			dayStart.add(1, 'day');
+			dayIndex++;
+
+			// detect when span won't go fully into the next day,
+			// and mutate the latest seg to the be the end.
+			if (
+				seg && !seg.isEnd && span.end.hasTime() &&
+				span.end < dayStart.clone().add(this.view.nextDayThreshold)
+			) {
+				seg.end = span.end.clone();
+				seg.isEnd = true;
+				break;
+			}
 		}
 
 		return segs;
@@ -123,11 +138,12 @@ var ListViewGrid = Grid.extend({
 
 		if (!segs.length) {
 			this.renderEmptyMessage();
-			return segs;
 		}
 		else {
-			return this.renderSegList(segs);
+			this.renderSegList(segs);
 		}
+
+		return segs;
 	},
 
 	renderEmptyMessage: function() {
@@ -142,30 +158,47 @@ var ListViewGrid = Grid.extend({
 		);
 	},
 
-	// render the event segments in the view. returns the mutated array.
-	renderSegList: function(segs) {
+	// render the event segments in the view
+	renderSegList: function(allSegs) {
+		var segsByDay = this.groupSegsByDay(allSegs); // sparse array
+		var dayIndex;
+		var daySegs;
+		var i;
 		var tableEl = $('<table class="fc-list-table"><tbody/></table>');
 		var tbodyEl = tableEl.find('tbody');
-		var i, seg;
-		var dayDate;
 
-		this.sortEventSegs(segs);
+		for (dayIndex = 0; dayIndex < segsByDay.length; dayIndex++) {
+			daySegs = segsByDay[dayIndex];
+			if (daySegs) { // sparse array, so might be undefined
 
-		for (i = 0; i < segs.length; i++) {
-			seg = segs[i];
+				// append a day header
+				tbodyEl.append(this.dayHeaderHtml(
+					this.view.start.clone().add(dayIndex, 'days')
+				));
 
-			// append a day header
-			if (!dayDate || !seg.start.isSame(dayDate, 'day')) {
-				dayDate = seg.start.clone().stripTime();
-				tbodyEl.append(this.dayHeaderHtml(dayDate));
+				this.sortEventSegs(daySegs);
+
+				for (i = 0; i < daySegs.length; i++) {
+					tbodyEl.append(daySegs[i].el); // append event row
+				}
 			}
-
-			tbodyEl.append(seg.el); // append event row
 		}
 
 		this.el.empty().append(tableEl);
+	},
 
-		return segs; // return the sorted list
+	// Returns a sparse array of arrays, segs grouped by their dayIndex
+	groupSegsByDay: function(segs) {
+		var segsByDay = []; // sparse array
+		var i, seg;
+
+		for (i = 0; i < segs.length; i++) {
+			seg = segs[i];
+			(segsByDay[seg.dayIndex] || (segsByDay[seg.dayIndex] = []))
+				.push(seg);
+		}
+
+		return segsByDay;
 	},
 
 	// generates the HTML for the day headers that live amongst the event rows
@@ -203,13 +236,20 @@ var ListViewGrid = Grid.extend({
 		var url = event.url;
 		var timeHtml;
 
-		if (!seg.start.hasTime()) {
-			if (this.displayEventTime) {
+		if (event.allDay) {
+			timeHtml = view.getAllDayHtml();
+		}
+		else if (view.isMultiDayEvent(event)) { // if the event appears to span more than one day
+			if (seg.isStart || seg.isEnd) { // outer segment that probably lasts part of the day
+				timeHtml = htmlEscape(this.getEventTimeText(seg));
+			}
+			else { // inner segment that lasts the whole day
 				timeHtml = view.getAllDayHtml();
 			}
 		}
 		else {
-			timeHtml = htmlEscape(this.getEventTimeText(event)); // might return empty
+			// Display the normal time text for the *event's* times
+			timeHtml = htmlEscape(this.getEventTimeText(event));
 		}
 
 		if (url) {
@@ -217,9 +257,9 @@ var ListViewGrid = Grid.extend({
 		}
 
 		return '<tr class="' + classes.join(' ') + '">' +
-			(timeHtml ?
+			(this.displayEventTime ?
 				'<td class="fc-list-item-time ' + view.widgetContentClass + '">' +
-					timeHtml +
+					(timeHtml || '') +
 				'</td>' :
 				'') +
 			'<td class="fc-list-item-marker ' + view.widgetContentClass + '">' +
@@ -231,7 +271,7 @@ var ListViewGrid = Grid.extend({
 			'</td>' +
 			'<td class="fc-list-item-title ' + view.widgetContentClass + '">' +
 				'<a' + (url ? ' href="' + htmlEscape(url) + '"' : '') + '>' +
-					htmlEscape(seg.event.title) +
+					htmlEscape(seg.event.title || '') +
 				'</a>' +
 			'</td>' +
 		'</tr>';